Add followUser hook
